refactor(ios): extract IOSThreadStore to its own file
Move IOSThread, PersistedThread, and IOSThreadStore from
ThreadListView.swift to App/IOSThreadStore.swift. ThreadListView.swift
now contains only the SwiftUI view code (~200 lines vs 513 before).

Pure file reorganization — no logic changes.
